refactor(pipeline): extract decompose_issue + review_pr — 1000+ LOC into stages/
The two largest remaining stage handlers extracted:
- stages/decompose.py (557 LOC): Quartermaster issue decomposition with
  _triage_issue, _enumerable_ruff, _enumerable_mypy, _needs_further_decomposition,
  and the recursive decompose_issue function
- stages/review_pr.py (500 LOC): Gatekeeper PR review with approve/reject
  verdicts, plus auditor_review (verdict parsing + LLM call)

RuntimePipeline delegations restored for methods that were accidentally
removed during the bulk extraction: _count_passing, _count_failing,
_parse_violation_files, _extract_files_from_issue_body, auditor_review.

__init__.py: 1003 lines (from 1996 — halved in one commit).
760 tests pass, 0 failures.
