+++ src/stronghold/builders/pipeline/__init__.py
@@ -962,1018 +962,11 @@ class RuntimePipeline:
             return stripped
         raise ExtractionError("Could not extract HTML from response")
 
-    # ── Quartermaster: Issue Decomposition ────────────────────────────
 
     @staticmethod
-    def _triage_issue(title: str, body: str) -> str:
-        """Classify an issue into a decomposition strategy.
-
-        Returns one of:
-        - "atomic"          — single file, no decomposition needed
-        - "enumerable:ruff" — work is enumerable from ruff output
-        - "enumerable:mypy" — work is enumerable from mypy output
-        - "agentic"         — needs LLM-driven planning
-        """
-        import re as _re
-
-        text = f"{title}\n{body}".lower()
-
-        # Enumerable: tool-driven cleanup
-        if "ruff check" in text or "ruff errors" in text or "ruff lint" in text:
-            return "enumerable:ruff"
-        if "mypy --strict" in text or "mypy errors" in text or "type errors" in text:
-            return "enumerable:mypy"
-
-        # Atomic: single file mention with few criteria, no multi-file markers
-        path_matches = _re.findall(r"src/stronghold/[\w/]+\.(?:py|html)", text)
-        unique_paths = set(path_matches)
-        criteria_count = body.count("- [ ]") + body.count("- [x]")
-
-        # Strong multi-file signals — anything with these is NOT atomic
-        multi_file_markers = (
-            "files to create",
-            "files to modify",
-            "files to add",
-            "## files",
-            "new files",
-            ".github/workflows/",
-        )
-        has_multi_file_marker = any(m in text for m in multi_file_markers)
-
-        if (
-            len(unique_paths) <= 1
-            and criteria_count <= 3
-            and not has_multi_file_marker
-        ):
-            return "atomic"
-
-        # Otherwise, agentic LLM decomposition
-        return "agentic"
-
-    async def _enumerable_ruff(
-        self, run: Any, ws: str,
-    ) -> list[dict[str, Any]]:
-        """Run ruff and produce one step per file with errors.
-
-        Writes ruff JSON to a temp file (shell tool truncates stdout at
-        3000 chars; ruff output for many errors is much bigger). Then
-        reads the file and parses.
-        """
-        import json as _json
-        from collections import defaultdict
-
-        # Write ruff JSON to a temp file in the worktree
-        await self._td.execute(
-            "shell",
-            {
-                "command": (
-                    "ruff check src/stronghold/ "
-                    "--output-format=json --no-fix > .ruff_errors.json 2>&1 || true"
-                ),
-                "workspace": ws,
-            },
-        )
-
-        stdout = await self._read_file(".ruff_errors.json", ws)
-        if not stdout:
-            return []
-
-        # Find the JSON array in stdout (ruff outputs an array of objects)
-        try:
-            json_start = stdout.find("[")
-            json_end = stdout.rfind("]") + 1
-            if json_start == -1 or json_end == 0:
-                return []
-            ruff_errors = _json.loads(stdout[json_start:json_end])
-        except Exception:
-            return []
-
-        if not ruff_errors:
-            return []
-
-        # Group by file
-        by_file: dict[str, list[dict[str, Any]]] = defaultdict(list)
-        for err in ruff_errors:
-            fname = err.get("filename", "")
-            if fname.startswith("/"):
-                # Make path relative to workspace
-                fname = fname.split("/src/stronghold/", 1)[-1]
-                fname = "src/stronghold/" + fname
-            by_file[fname].append(err)
-
-        # Build steps — one per file
-        steps = []
-        for fname, errors in sorted(by_file.items()):
-            rules = sorted({e.get("code", "?") for e in errors})
-            error_lines = []
-            for e in errors[:20]:
-                code = e.get("code", "?")
-                msg = e.get("message", "")
-                loc = e.get("location", {})
-                line = loc.get("row", 0)
-                col = loc.get("column", 0)
-                fix_avail = "✓" if e.get("fix") else " "
-                error_lines.append(f"  {fix_avail} {code} {fname}:{line}:{col}  {msg}")
-            if len(errors) > 20:
-                error_lines.append(f"  ... and {len(errors) - 20} more")
-
-            body = (
-                f"## Description\n"
-                f"Fix all ruff violations in `{fname}`.\n\n"
-                f"## Errors ({len(errors)} total)\n"
-                f"Rules: {', '.join(rules)}\n\n"
-                f"```\n{chr(10).join(error_lines)}\n```\n\n"
-                f"## Acceptance Criteria\n"
-                f"- [ ] `ruff check {fname}` returns zero errors\n"
-                f"- [ ] `ruff format --check {fname}` passes\n"
-                f"- [ ] No functional changes (lint/format only)\n\n"
-                f"## Implementation Notes\n"
-                f"Run `ruff check --fix {fname}` and `ruff format {fname}` "
-                f"first to handle auto-fixable items, then manually fix the rest.\n\n"
-                f"## Files\n- {fname}"
-            )
-            steps.append({
-                "title": (
-                    f"fix: ruff cleanup in "
-                    f"{fname.replace('src/stronghold/', '')}"
-                ),
-                "body": body,
-                "depends_on": [],
-                "file": fname,
-            })
-
-        return steps
-
-    async def _enumerable_mypy(
-        self, run: Any, ws: str,
-    ) -> list[dict[str, Any]]:
-        """Run mypy and produce one step per file with errors."""
-        from collections import defaultdict
-        import re as _re
-
-        await self._td.execute(
-            "shell",
-            {
-                "command": (
-                    "mypy src/stronghold/ --strict --no-error-summary "
-                    "> .mypy_errors.txt 2>&1 || true"
-                ),
-                "workspace": ws,
-            },
-        )
-
-        stdout = await self._read_file(".mypy_errors.txt", ws)
-        if not stdout:
-            return []
-
-        # Parse mypy lines: path:line: error: message  [code]
-        line_pat = _re.compile(
-            r"^(src/stronghold/[^:]+):(\d+):(?:\d+:)?\s*(error|warning):\s*(.+?)(?:\s+\[([^\]]+)\])?$",
-            _re.MULTILINE,
-        )
-        by_file: dict[str, list[dict[str, str]]] = defaultdict(list)
-        for m in line_pat.finditer(stdout):
-            fname, line, _sev, msg, code = m.groups()
-            by_file[fname].append({
-                "line": line,
-                "message": msg.strip(),
-                "code": code or "",
-            })
-
-        if not by_file:
-            return []
-
-        steps = []
-        for fname, errors in sorted(by_file.items()):
-            error_lines = "\n".join(
-                f"  {e['line']}: {e['message']}"
-                + (f"  [{e['code']}]" if e["code"] else "")
-                for e in errors[:30]
-            )
-            if len(errors) > 30:
-                error_lines += f"\n  ... and {len(errors) - 30} more"
-
-            body = (
-                f"## Description\n"
-                f"Fix mypy --strict errors in `{fname}`.\n\n"
-                f"## Errors ({len(errors)} total)\n"
-                f"```\n{error_lines}\n```\n\n"
-                f"## Acceptance Criteria\n"
-                f"- [ ] `mypy {fname} --strict` passes with zero errors\n"
-                f"- [ ] No type-ignore comments added unless absolutely necessary\n"
-                f"- [ ] Existing behavior preserved\n\n"
-                f"## Files\n- {fname}"
-            )
-            steps.append({
-                "title": (
-                    f"fix: mypy strict in "
-                    f"{fname.replace('src/stronghold/', '')}"
-                ),
-                "body": body,
-                "depends_on": [],
-                "file": fname,
-            })
-
-        return steps
-
-    @staticmethod
-    def _needs_further_decomposition(title: str, body: str) -> bool:
-        """Heuristic: is this sub-issue still too broad for Mason to solve?
-
-        Signals:
-        - Mentions multiple distinct directories under src/stronghold/
-        - Mentions "across repo", "whole repo", "all files"
-        - Has more than 3 acceptance criteria touching different files
-        - Title prefixed with "cleanup", "refactor all", "fix all"
-        """
-        import re as _re
-
-        text = f"{title}\n{body}".lower()
-
-        # Broad-scope keywords
-        broad_keywords = [
-            "across repo", "across the repo", "whole repo",
-            "all files", "every file", "entire codebase",
-        ]
-        if any(kw in text for kw in broad_keywords):
-            return True
-
-        # Title prefixes indicating wide scope
-        broad_prefixes = ("cleanup", "refactor all", "fix all", "migrate all")
-        if any(title.lower().strip().startswith(p) for p in broad_prefixes):
-            return True
-
-        # Count distinct directories mentioned
-        paths = _re.findall(r"src/stronghold/([\w]+(?:/[\w]+)*)", text)
-        directories: set[str] = set()
-        for p in paths:
-            # Take first 2 path segments as "directory"
-            parts = p.split("/")
-            directories.add("/".join(parts[:2]))
-        if len(directories) > 2:
-            return True
-
-        return False
-
-    async def decompose_issue(
-        self, run: Any, *, depth: int = 0, max_depth: int = 3,
-    ) -> StageResult:
-        """Quartermaster: decompose a parent issue into sub-issues with dependencies.
-
-        If depth < max_depth, children that are still too broad will be
-        recursively decomposed. Parent issues get an 'epic' label so the
-        scheduler skips them and works the leaves instead.
-
-        1. Read parent issue + repo context
-        2. LLM produces JSON with steps + depends_on (local indices)
-        3. For each step: create_issue, then create_sub_issue(parent, child)
-        4. After all created: add_blocked_by edges per the depends_on map
-        5. For each child still too broad: recurse (up to max_depth)
-        6. Label parent as 'epic' so the scheduler skips it
-        7. Post a summary comment on the parent issue
-        """
-        import json as _json
-
-        owner, repo = run.repo.split("/")
-        ws = getattr(run, "_workspace_path", "")
-        issue_content = getattr(run, "_issue_content", "")
-        issue_title = getattr(run, "_issue_title", "")
-
-        # ── Triage: pick the right strategy ──
-        strategy = self._triage_issue(issue_title, issue_content)
-        logger.info(
-            "Quartermaster triage on #%s → %s",
-            run.issue_number, strategy,
-        )
-
-        plan_summary = ""
-        steps: list[dict[str, Any]] = []
-
-        if strategy == "atomic":
-            # No decomposition needed — leave the issue for Mason to work directly
-            return StageResult(
-                success=True,
-                summary=(
-                    f"Triage: atomic — single-file work order, "
-                    f"no decomposition needed."
-                ),
-                evidence={
-                    "parent": run.issue_number,
-                    "strategy": "atomic",
-                    "depth": depth,
-                    "created": [],
-                },
-            )
-
-        elif strategy == "enumerable:ruff":
-            steps = await self._enumerable_ruff(run, ws)
-            plan_summary = (
-                f"Enumerable ruff decomposition: {len(steps)} files with errors. "
-                f"One sub-issue per file."
-            )
-
-        elif strategy == "enumerable:mypy":
-            steps = await self._enumerable_mypy(run, ws)
-            plan_summary = (
-                f"Enumerable mypy decomposition: {len(steps)} files with errors. "
-                f"One sub-issue per file."
-            )
-
-        else:  # agentic
-            file_listing = await self._list_files("src/", ws)
-
-            relevant_files = ""
-            import re as _re
-            mentioned_paths = _re.findall(
-                r"(src/stronghold/[\w/]+\.py)", issue_content,
-            )
-            for path in mentioned_paths[:5]:
-                content = await self._read_file(path, ws)
-                if content:
-                    relevant_files += f"\n# --- {path} ---\n{content[:2000]}\n"
-
-            template = await self._get_prompt("builders.quartermaster.decompose")
-            if not template:
-                from stronghold.builders.prompts import QUARTERMASTER_DECOMPOSE
-                template = QUARTERMASTER_DECOMPOSE
-
-            prompt = self._render(
-                template,
-                issue_number=str(run.issue_number),
-                issue_title=issue_title,
-                issue_content=issue_content,
-                file_listing=file_listing[:4000],
-                relevant_files=relevant_files[:6000],
-            )
-
-            plan = await self._llm_extract(
-                prompt, self._frank_model, extract_json, "decomposition plan",
-            )
-            steps = plan.get("steps", [])
-            plan_summary = plan.get("summary", "")
-
-        if not steps:
-            return StageResult(
-                success=False,
-                summary=f"Triage: {strategy} produced no steps",
-            )
-
-        # Enumerable strategies are uncapped; agentic capped at 25.
-        # The cap exists because LLM decompositions tend to either over-split
-        # (50 trivial steps) or under-split (3 huge steps); the cap is a
-        # heuristic ceiling that says "if you produced more than this, your
-        # parent epic is genuinely too broad and a human should split it
-        # first." 25 fits real-world v0.9 epics like 'populate priority_tier
-        # in 15+ agent.yaml files'; the prior 10 was rejecting them as
-        # 'too broad' when they were actually correctly enumerated.
-        AGENTIC_CAP = 25
-        if strategy == "agentic" and len(steps) > AGENTIC_CAP:
-            return StageResult(
-                success=False,
-                summary=(
-                    f"Too many steps ({len(steps)} > {AGENTIC_CAP}) — "
-                    f"parent is too broad. Split it into narrower epics first."
-                ),
-            )
-
-        # Hard safety cap to prevent runaway issue creation
-        if len(steps) > 100:
-            steps = steps[:100]
-            plan_summary += f" (capped at 100 from larger set)"
-
-        # Create all child issues first, recording step index → issue number
-        created: list[dict[str, Any]] = []
-        for i, step in enumerate(steps):
-            title = step.get("title", f"sub-issue {i + 1}")
-            body = step.get("body", "")
-            body += f"\n\n---\n_Sub-issue of #{run.issue_number}, step {i + 1} of {len(steps)}_"
-
-            result = await self._td.execute(
-                "github",
-                {
-                    "action": "create_issue",
-                    "owner": owner, "repo": repo,
-                    "title": title, "body": body,
-                    "labels": ["builders", "quartermaster"],
-                },
-            )
-            if result.startswith("Error:"):
-                logger.error("Failed to create sub-issue %d: %s", i, result)
-                return StageResult(
-                    success=False,
-                    summary=f"Failed to create sub-issue {i + 1}: {result[:200]}",
-                )
-            try:
-                data = _json.loads(result)
-            except Exception:
-                return StageResult(
-                    success=False, summary=f"Bad create_issue response: {result[:200]}",
-                )
-            created.append({
-                "index": i,
-                "number": data["number"],
-                "title": title,
-                "depends_on": step.get("depends_on", []),
-            })
-
-        # Link each child as a sub-issue of the parent
-        for c in created:
-            await self._td.execute(
-                "github",
-                {
-                    "action": "create_sub_issue",
-                    "owner": owner, "repo": repo,
-                    "issue_number": run.issue_number,
-                    "sub_issue_number": c["number"],
-                },
-            )
-
-        # Add blocked_by edges per depends_on (local index → issue number)
-        for c in created:
-            for dep_idx in c["depends_on"]:
-                if not isinstance(dep_idx, int) or dep_idx >= len(created):
-                    continue
-                blocker = created[dep_idx]
-                await self._td.execute(
-                    "github",
-                    {
-                        "action": "add_blocked_by",
-                        "owner": owner, "repo": repo,
-                        "issue_number": c["number"],
-                        "blocker_issue_number": blocker["number"],
-                    },
-                )
-
-        # Recurse on children that are still too broad
-        recursive_depth = depth + 1
-        recurse_counts: list[dict[str, Any]] = []
-        if recursive_depth <= max_depth:
-            from types import SimpleNamespace
-
-            for c in created:
-                # Get the step body that was used to create this child
-                step_body = ""
-                for i, step in enumerate(steps):
-                    if i == c["index"]:
-                        step_body = step.get("body", "")
-                        break
-
-                if not self._needs_further_decomposition(c["title"], step_body):
-                    continue
-
-                logger.info(
-                    "Quartermaster recursing on #%d (depth %d)",
-                    c["number"], recursive_depth,
-                )
-                child_run = SimpleNamespace(
-                    run_id=f"qm-{recursive_depth}-{c['number']}",
-                    issue_number=c["number"],
-                    repo=run.repo,
-                    _issue_title=c["title"],
-                    _issue_content=step_body,
-                    _workspace_path=ws,
-                )
-                child_result = await self.decompose_issue(
-                    child_run, depth=recursive_depth, max_depth=max_depth,
-                )
-                recurse_counts.append({
-                    "parent": c["number"],
-                    "success": child_result.success,
-                    "sub_created": (
-                        child_result.evidence.get("created", [])
-                        if child_result.success
-                        else []
-                    ),
-                })
-
-        # Label this parent as 'epic' so the scheduler skips it
-        await self._td.execute(
-            "github",
-            {
-                "action": "add_labels",
-                "owner": owner, "repo": repo,
-                "issue_number": run.issue_number,
-                "labels": ["epic"],
-            },
-        )
-
-        # Post summary comment to the parent
-        lines = [
-            f"## Quartermaster Decomposition (depth {depth}, strategy: {strategy})\n"
-        ]
-        lines.append(f"{plan_summary}\n")
-        lines.append(f"**{len(created)} sub-issues created:**\n")
-        for c in created:
-            deps = c["depends_on"]
-            dep_str = ""
-            if deps:
-                dep_numbers = [f"#{created[d]['number']}" for d in deps if d < len(created)]
-                dep_str = f" _(blocked by {', '.join(dep_numbers)})_"
-            # Mark recursed children
-            recursed = any(r["parent"] == c["number"] for r in recurse_counts)
-            marker = " 🔻 _(further decomposed)_" if recursed else ""
-            lines.append(f"- #{c['number']} {c['title']}{dep_str}{marker}")
-
-        if recurse_counts:
-            total_leaves = sum(len(r["sub_created"]) for r in recurse_counts)
-            lines.append(
-                f"\n_Recursed on {len(recurse_counts)} children, "
-                f"created {total_leaves} leaf sub-issues._"
-            )
-
-        summary_text = "\n".join(lines)
-
-        await self._td.execute(
-            "github",
-            {
-                "action": "post_pr_comment",
-                "owner": owner, "repo": repo,
-                "issue_number": run.issue_number,
-                "body": summary_text,
-            },
-        )
-
-        return StageResult(
-            success=True,
-            summary=summary_text,
-            evidence={
-                "parent": run.issue_number,
-                "depth": depth,
-                "strategy": strategy,
-                "created": [{"number": c["number"], "title": c["title"]} for c in created],
-                "dependency_count": sum(len(c["depends_on"]) for c in created),
-                "recursed": recurse_counts,
-            },
-        )
-
-    # ── Gatekeeper: PR Review ────────────────────────────────────────
-
-    async def review_pr(
-        self,
-        *,
-        owner: str,
-        repo: str,
-        pr_number: int,
-        auto_merge_enabled: bool = False,
-        allowed_authors: tuple[str, ...] = (),
-        coverage_tolerance_pct: float = -1.0,
-        protected_branches: tuple[str, ...] = ("main", "master"),
-    ) -> StageResult:
-        """Gatekeeper: review a PR end-to-end and either approve or request changes.
-
-        Phases:
-        1. Intake: fetch PR metadata, diff, parent issue
-        2. Scope: read full changed files + siblings + callers
-        3. Mechanical: ruff, mypy, bandit, pytest on changed files
-        4. LLM semantic review: feeds context to Opus, extracts JSON verdict
-        5. Act: post review (APPROVE or REQUEST_CHANGES), merge if approved
-        """
-        import json as _json
-
-        td = self._td
-
-        # ── Phase 1: Intake ────────────────────────────────────────
-        pr_raw = await td.execute(
-            "github",
-            {
-                "action": "get_pr",
-                "owner": owner, "repo": repo,
-                "issue_number": pr_number,
-            },
-        )
-        if pr_raw.startswith("Error:"):
-            return StageResult(
-                success=False, summary=f"Cannot fetch PR: {pr_raw[:200]}",
-            )
-        try:
-            pr = _json.loads(pr_raw)
-        except Exception:
-            return StageResult(
-                success=False, summary="Bad PR response",
-            )
-
-        # Identify parent issue from PR title (pattern: "feat: #NNN —")
-        import re as _re
-        issue_match = _re.search(r"#(\d+)", pr.get("title", ""))
-        parent_issue_number = int(issue_match.group(1)) if issue_match else None
-        issue_body = ""
-        if parent_issue_number:
-            issue_raw = await td.execute(
-                "github",
-                {
-                    "action": "get_issue",
-                    "owner": owner, "repo": repo,
-                    "issue_number": parent_issue_number,
-                },
-            )
-            if not issue_raw.startswith("Error:"):
-                try:
-                    issue_body = _json.loads(issue_raw).get("body", "") or ""
-                except Exception:
-                    issue_body = ""
-
-        # Fetch files changed
-        files_raw = await td.execute(
-            "github",
-            {
-                "action": "list_pr_files",
-                "owner": owner, "repo": repo,
-                "issue_number": pr_number,
-            },
-        )
-        try:
-            pr_files = _json.loads(files_raw) if not files_raw.startswith("Error:") else []
-        except Exception:
-            pr_files = []
-
-        changed_paths = [f["filename"] for f in pr_files if f["filename"].startswith("src/")]
-        py_files = [p for p in changed_paths if p.endswith(".py")]
-
-        # ── Phase 2: Scope (need a worktree) ──────────────────────
-        # Use workspace tool to ensure we have the PR branch checked out
-        ws_result = await td.execute(
-            "workspace",
-            {
-                "action": "create",
-                "issue_number": parent_issue_number or pr_number,
-                "owner": owner, "repo": repo,
-            },
-        )
-        ws_path = ""
-        if not ws_result.startswith("Error:"):
-            try:
-                ws_path = _json.loads(ws_result).get("path", "")
-            except Exception:
-                ws_path = ""
-
-        # Checkout the PR head branch in the workspace
-        if ws_path and pr.get("head", {}).get("ref"):
-            head_ref = pr["head"]["ref"]
-            await td.execute(
-                "shell",
-                {
-                    "command": f"git fetch origin {head_ref} && "
-                    f"git checkout -B {head_ref} origin/{head_ref}",
-                    "workspace": ws_path,
-                },
-            )
-
-        # Read full changed files
-        changed_files_content = ""
-        for fpath in changed_paths[:10]:
-            content = await self._read_file(fpath, ws_path)
-            if content:
-                changed_files_content += (
-                    f"\n# --- {fpath} ---\n{content[:3000]}\n"
-                )
-
-        # Read siblings of new files for parallel structure
-        sibling_files_content = ""
-        for fpath in changed_paths[:5]:
-            # Only for newly added files
-            added = any(
-                f["filename"] == fpath and f["status"] == "added"
-                for f in pr_files
-            )
-            if not added:
-                continue
-            import os as _os
-            dir_path = _os.path.dirname(fpath)
-            if not dir_path:
-                continue
-            siblings_raw = await td.execute(
-                "glob_files",
-                {
-                    "pattern": f"{dir_path}/*.py",
-                    "workspace": ws_path,
-                    "max_results": 5,
-                },
-            )
-            if siblings_raw.startswith("Error:"):
-                continue
-            try:
-                siblings_data = _json.loads(siblings_raw)
-                sibling_paths = siblings_data.get("files", [])
-            except Exception:
-                sibling_paths = []
-            for sib in sibling_paths[:3]:
-                if sib == fpath:
-                    continue
-                sib_content = await self._read_file(sib, ws_path)
-                if sib_content:
-                    sibling_files_content += (
-                        f"\n# --- {sib} (sibling pattern) ---\n"
-                        f"{sib_content[:2000]}\n"
-                    )
-
-        # Callers of changed symbols — grep for imports
-        callers_content = ""
-        for fpath in py_files[:3]:
-            module = fpath.replace("src/", "").replace("/", ".").replace(".py", "")
-            callers_raw = await td.execute(
-                "grep_content",
-                {
-                    "pattern": f"from {module}|import {module}",
-                    "workspace": ws_path,
-                    "glob": "**/*.py",
-                    "max_results": 10,
-                },
-            )
-            if callers_raw and not callers_raw.startswith("Error:"):
-                try:
-                    callers_data = _json.loads(callers_raw)
-                    matches = callers_data.get("matches", [])
-                    for m in matches[:5]:
-                        callers_content += (
-                            f"- {m['file']}:{m['line']}: {m['content']}\n"
-                        )
-                except Exception:
-                    pass
-
-        # Read CLAUDE.md and ONBOARDING.md
-        claude_md = (await self._read_file("CLAUDE.md", ws_path))[:4000]
-        onboarding_md = (await self._read_file("ONBOARDING.md", ws_path))[:4000]
-        repo_standards = (
-            f"## CLAUDE.md\n{claude_md}\n\n## ONBOARDING.md\n{onboarding_md}"
-        )
-
-        # ── Phase 3: Mechanical gates ──────────────────────────────
-        mechanical_results: dict[str, str] = {}
-        if py_files:
-            files_str = " ".join(py_files)
-            for gate_name, cmd in [
-                ("ruff_check", f"ruff check {files_str}"),
-                ("ruff_format", f"ruff format --check {files_str}"),
-                ("mypy", f"mypy {files_str} --strict"),
-                ("bandit", f"bandit {files_str} -ll"),
-            ]:
-                result = await td.execute(
-                    "shell",
-                    {"command": cmd, "workspace": ws_path},
-                )
-                mechanical_results[gate_name] = result[:500]
-
-        mechanical_summary = "\n".join(
-            f"{k}: {v[:200]}" for k, v in mechanical_results.items()
-        )
-        mechanical_pass = all(
-            '"passed": true' in v or v.startswith("Error:") or "Success" in v
-            for v in mechanical_results.values()
-        )
-
-        # ── Coverage check ─────────────────────────────────────────
-        coverage_summary = "Coverage check skipped (no baseline available)"
-        if py_files and ws_path:
-            cov_modules = ",".join(
-                f.replace("src/", "").replace("/", ".").replace(".py", "")
-                for f in py_files
-            )
-            cov_result = await td.execute(
-                "shell",
-                {
-                    "command": (
-                        f"python -m pytest tests/ -q --cov={cov_modules} "
-                        f"--cov-report=term --no-header 2>&1 | tail -20"
-                    ),
-                    "workspace": ws_path,
-                },
-            )
-            if not cov_result.startswith("Error:"):
-                coverage_summary = f"```\n{cov_result[:1500]}\n```"
-
-        # ── Phase 4: LLM semantic review ───────────────────────────
-        template = await self._get_prompt("builders.gatekeeper.review_pr")
-        if not template:
-            from stronghold.builders.prompts import GATEKEEPER_REVIEW_PR
-            template = GATEKEEPER_REVIEW_PR
-
-        prompt = self._render(
-            template,
-            pr_number=str(pr_number),
-            pr_title=pr.get("title", ""),
-            pr_author=pr.get("user", ""),
-            pr_body=(pr.get("body") or "")[:2000],
-            base_branch=pr.get("base", {}).get("ref", "main"),
-            head_branch=pr.get("head", {}).get("ref", ""),
-            files_count=str(pr.get("changed_files", len(pr_files))),
-            additions=str(pr.get("additions", 0)),
-            deletions=str(pr.get("deletions", 0)),
-            issue_number=str(parent_issue_number or ""),
-            issue_body=issue_body[:3000],
-            mechanical_result=mechanical_summary[:2000],
-            coverage_summary=coverage_summary[:1500],
-            changed_files=changed_files_content[:15000],
-            sibling_files=sibling_files_content[:6000],
-            callers=callers_content[:3000],
-            repo_standards=repo_standards[:8000],
-        )
-
-        verdict = await self._llm_extract(
-            prompt, self._mason_model, extract_json, "gatekeeper verdict",
-        )
-
-        decision = verdict.get("decision", "request_changes")
-        summary = verdict.get("summary", "")
-        blockers = verdict.get("blockers", [])
-        checked = verdict.get("checked", [])
-        suggestions = verdict.get("suggestions", [])
-
-        # Mechanical failures force request_changes
-        if not mechanical_pass and decision == "approve":
-            decision = "request_changes"
-            blockers.append({
-                "file": "(multiple)",
-                "line": 0,
-                "severity": "error",
-                "category": "mechanical",
-                "message": f"Mechanical gates failed:\n{mechanical_summary[:1000]}",
-            })
-
-        # ── Phase 5: Act ───────────────────────────────────────────
-        review_body_lines = [
-            f"## Gatekeeper Review",
-            "",
-            f"**Decision:** {decision.upper()}",
-            f"**Summary:** {summary}",
-            "",
-        ]
-        if checked:
-            review_body_lines.append("### What I checked")
-            for c in checked[:20]:
-                review_body_lines.append(f"- ✓ {c}")
-            review_body_lines.append("")
-
-        if blockers:
-            review_body_lines.append("### Blockers")
-            for b in blockers[:20]:
-                line_ref = f":{b.get('line')}" if b.get("line") else ""
-                review_body_lines.append(
-                    f"- **{b.get('category', '?')}** "
-                    f"`{b.get('file', '?')}{line_ref}` — "
-                    f"{b.get('message', '')}"
-                )
-            review_body_lines.append("")
-
-        if suggestions:
-            review_body_lines.append("### Suggestions (non-blocking)")
-            for s in suggestions[:10]:
-                line_ref = f":{s.get('line')}" if s.get("line") else ""
-                review_body_lines.append(
-                    f"- `{s.get('file', '?')}{line_ref}` — "
-                    f"{s.get('message', '')}"
-                )
-
-        review_body = "\n".join(review_body_lines)
-
-        event = "APPROVE" if decision == "approve" else "REQUEST_CHANGES"
-
-        review_result = await td.execute(
-            "github",
-            {
-                "action": "review_pr",
-                "owner": owner, "repo": repo,
-                "issue_number": pr_number,
-                "event": event,
-                "body": review_body,
-            },
-        )
-
-        merged = False
-        merge_message = ""
-        if decision == "approve" and auto_merge_enabled:
-            # Guardrails
-            author_ok = (
-                not allowed_authors
-                or pr.get("user", "") in allowed_authors
-            )
-            branch_ok = pr.get("base", {}).get("ref") not in protected_branches
-            if author_ok and branch_ok:
-                merge_raw = await td.execute(
-                    "github",
-                    {
-                        "action": "merge_pr",
-                        "owner": owner, "repo": repo,
-                        "issue_number": pr_number,
-                        "merge_method": "squash",
-                        "commit_title": pr.get("title", f"PR #{pr_number}"),
-                    },
-                )
-                if not merge_raw.startswith("Error:"):
-                    try:
-                        merge_data = _json.loads(merge_raw)
-                        merged = bool(merge_data.get("merged", False))
-                        merge_message = merge_data.get("message", "")
-                    except Exception:
-                        pass
-            else:
-                merge_message = (
-                    f"auto_merge guardrails rejected: "
-                    f"author_ok={author_ok} branch_ok={branch_ok}"
-                )
-
-        # Post verdict to the parent issue (if there is one)
-        if parent_issue_number:
-            parent_body = (
-                f"## Gatekeeper Verdict on PR #{pr_number}\n\n"
-                f"**Decision:** {decision.upper()}\n"
-                f"{summary}\n"
-            )
-            if merged:
-                parent_body += f"\n**Merged.** {merge_message}\n"
-            elif merge_message:
-                parent_body += f"\n_{merge_message}_\n"
-            await td.execute(
-                "github",
-                {
-                    "action": "post_pr_comment",
-                    "owner": owner, "repo": repo,
-                    "issue_number": parent_issue_number,
-                    "body": parent_body,
-                },
-            )
-
-        return StageResult(
-            success=(decision == "approve"),
-            summary=f"{event}: {summary}",
-            evidence={
-                "pr": pr_number,
-                "decision": decision,
-                "blockers_count": len(blockers),
-                "blockers": blockers[:10],
-                "checked_count": len(checked),
-                "merged": merged,
-                "merge_message": merge_message,
-                "mechanical_pass": mechanical_pass,
-            },
-        )
-
-    # ── Auditor Review ───────────────────────────────────────────────
-
-    async def auditor_review(
-        self,
-        stage: str,
-        evidence: dict[str, Any],
-    ) -> tuple[bool, str]:
-        """Auditor reviews concrete evidence using composed prompts from the library."""
-        # Get stage-specific context from prompt library
-        stage_context = await self._get_prompt(f"builders.auditor.stage.{stage}")
-        if not stage_context:
-            # Fallback to hardcoded context dict
-            ctx = _AUDITOR_STAGE_CONTEXT.get(stage, {})
-            purpose = ctx.get("purpose", "Complete the stage")
-            scope = ctx.get("scope", "")
-            out_of_scope = ctx.get("out_of_scope", "")
-            checklist = ctx.get("checklist", [])
-            rejection_format = ctx.get("rejection_format", "Be specific")
-            checklist_text = "\n".join(f"- [ ] {item}" for item in checklist)
-        else:
-            # Parse structured context from prompt library
-            import yaml
-            try:
-                ctx = yaml.safe_load(stage_context)
-            except Exception:
-                ctx = {}
-            purpose = ctx.get("purpose", "Complete the stage")
-            scope = ctx.get("scope", "")
-            out_of_scope = ctx.get("out_of_scope", "")
-            checklist_raw = ctx.get("checklist", [])
-            rejection_format = ctx.get("rejection_format", "Be specific")
-            checklist_text = "\n".join(
-                f"- [ ] {item}" for item in (checklist_raw if isinstance(checklist_raw, list) else [])
-            )
-
-        evidence_text = "\n".join(
-            f"**{k}:**\n{v}" if isinstance(v, str) else f"**{k}:** {v}"
-            for k, v in evidence.items()
-        )
-
-        # Compose review prompt from library
-        review_template = await self._get_prompt("builders.auditor.review")
-        prompt = self._render(
-            review_template,
-            stage=stage,
-            purpose=purpose,
-            scope=scope,
-            out_of_scope=out_of_scope,
-            checklist=checklist_text,
-            evidence=evidence_text,
-            rejection_format=rejection_format,
-        )
-
-        text = await self._llm_call(prompt, self._auditor_model)
-        from stronghold.builders.pipeline.stages.auditor import parse_verdict
-
-        approved = parse_verdict(text)
-        auditor_logger.info(
-            "[AUDITOR] stage=%s approved=%s first80=%s",
-            stage, approved, text[:80] if text else "EMPTY",
-            extra={"run_id": "-"},
-        )
-        return approved, text
-
-    # ── Utilities ────────────────────────────────────────────────────
-
-    # ── Delegations to PytestRunner ─────────────────────────────────────
+    def _extract_files_from_issue_body(issue_body: str) -> list[str]:
+        from stronghold.builders.pipeline.github_helpers import extract_files_from_issue_body
+        return extract_files_from_issue_body(issue_body)
 
     @staticmethod
     def _count_passing(pytest_output: str) -> int:
@@ -1990,7 +983,26 @@ class RuntimePipeline:
         from stronghold.builders.pipeline.pytest_runner import PytestRunner
         return PytestRunner.parse_violation_files(output)
 
-    @staticmethod
-    def _extract_files_from_issue_body(issue_body: str) -> list[str]:
-        from stronghold.builders.pipeline.github_helpers import extract_files_from_issue_body
-        return extract_files_from_issue_body(issue_body)
+    # ── Auditor review (used by _execute_one_stage in routes) ──────────
+
+    async def auditor_review(self, stage: str, evidence: dict[str, Any]) -> tuple[bool, str]:
+        from stronghold.builders.pipeline.stages.review_pr import auditor_review
+        return await auditor_review(stage, evidence, pipeline=self)
+
+    # ── Quartermaster: Issue Decomposition ────────────────────────────
+
+    async def decompose_issue(
+        self, run: Any, feedback: str = "",
+        *, max_sub_issues: int = 25, parent_issue_number: int | None = None,
+    ) -> Any:
+        from stronghold.builders.pipeline.stages.decompose import decompose_issue
+        return await decompose_issue(
+            run, pipeline=self, feedback=feedback,
+            max_sub_issues=max_sub_issues, parent_issue_number=parent_issue_number,
+        )
+
+    # ── Gatekeeper: PR Review ────────────────────────────────────────
+
+    async def review_pr(self, *args: Any, **kwargs: Any) -> Any:
+        from stronghold.builders.pipeline.stages.review_pr import review_pr
+        return await review_pr(*args, pipeline=self, **kwargs)
